+++ core/src/com/giancola/battleship/actors/ActorExtensions.kt
@@ -1,13 +1,12 @@
 package com.giancola.battleship.actors
 
-import com.badlogic.gdx.math.Intersector
 import com.badlogic.gdx.math.Polygon
 import com.badlogic.gdx.math.Rectangle
 import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.Group
 import com.badlogic.gdx.scenes.scene2d.Stage
 
-fun Actor.getBounds(): Rectangle {
+fun Actor.getBoundingBox(): Rectangle {
     val vertices = floatArrayOf(0f, 0f, this.width, 0f, this.width, this.height, 0f, this.height)
     val boundaryPolygon = Polygon(vertices)
     boundaryPolygon.setPosition(this.x, this.y)
@@ -17,7 +16,7 @@ fun Actor.getBounds(): Rectangle {
     return boundaryPolygon.boundingRectangle
 }
 
-fun Actor.overlaps(other: Actor): Boolean  = this.getBounds().overlaps(other.getBounds())
+fun Actor.overlaps(other: Actor): Boolean  = this.getBoundingBox().overlaps(other.getBoundingBox())
 
 
 fun Actor.getDescendants(): List<Actor> {