+++ core/src/com/giancola/battleship/actors/CombatEnemyBoard.kt
@@ -4,14 +4,13 @@ import com.badlogic.gdx.scenes.scene2d.Group
 import com.giancola.battleship.GameConstants
 import com.giancola.battleship.GameConstants.TILE_SIZE
 import com.giancola.battleship.LayoutConstants
-import com.giancola.battleship.screens.PlacementScreen
+import com.giancola.battleship.screens.CombatScreen
 
 
-class PlayerBoard(val screen: PlacementScreen, val nRows: Int = GameConstants.N_ROWS, val nCols: Int = GameConstants.N_COLS): Group(), DropTarget {
-    override var targetable = true
+class CombatEnemyBoard(val screen: CombatScreen, val nRows: Int = GameConstants.N_ROWS, val nCols: Int = GameConstants.N_COLS): Group() {
 
     val tiles = Array(nRows) { i -> Array(nCols) { j ->
-        Tile(this, i, j).also {
+        Tile(i, j, TILE_SIZE).also {
             it.setPosition(TILE_SIZE * i, TILE_SIZE * j)
             it.setSize(TILE_SIZE, TILE_SIZE)
 
@@ -20,7 +19,7 @@ class PlayerBoard(val screen: PlacementScreen, val nRows: Int = GameConstants.N_
     } }
 
     init {
-        val boardLayout = LayoutConstants.standard2worldCoords(LayoutConstants.playerBoard)
+        val boardLayout = LayoutConstants.standard2worldCoords(LayoutConstants.combatEnemyBoard)
         this.setPosition(boardLayout.x, boardLayout.y)
         this.setSize(TILE_SIZE * nCols, TILE_SIZE * nRows)
     }