+++ core/src/com/giancola/battleship/actors/Gestures.kt
@@ -87,7 +87,7 @@ class ShipGestureListener(private val ship: ShipActor) : ActorGestureListener() 
             }
         }
 
-        if (dropTarget == null || (dropTarget !is PlayerBoard)) {
+        if (dropTarget == null || (dropTarget !is PlacementBoard)) {
             ship.addAction(Actions.moveTo(startPositionX, startPositionY,2 * movementAnimDuration, Interpolation.pow3))
             shipCallbacks.onDragCanceled()