+++ core/src/com/giancola/battleship/screens/DemoScreen.kt
@@ -3,12 +3,12 @@ package com.giancola.battleship.screens
 import com.badlogic.gdx.graphics.Color.*
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer.ShapeType.*
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer
-import com.giancola.battleship.App
+import com.giancola.battleship.BattleshipGame
 import ktx.app.KtxScreen
 import ktx.graphics.*
 
 
-class DemoScreen(val app: App) : KtxScreen {
+class DemoScreen(val battleshipGame: BattleshipGame) : KtxScreen {
     private val sr = ShapeRenderer()
     private var rotation = 0.0f
 
@@ -21,7 +21,7 @@ class DemoScreen(val app: App) : KtxScreen {
 
         with(sr) {
             use(Line) {
-                translate(app.width / 2, app.height / 2, 0f)
+                translate(battleshipGame.width / 2, battleshipGame.height / 2, 0f)
                 rotate(0f, 0f, 1f, rotation)
                 color = WHITE
                 rect(0f - 75, 0f - 75, 150f, 150f)
@@ -33,13 +33,13 @@ class DemoScreen(val app: App) : KtxScreen {
                 circle(0f, 0f, 200f, 25)
 
                 color = PINK
-                circle(0f, app.height, 200f, 25)
+                circle(0f, battleshipGame.height, 200f, 25)
 
                 color = YELLOW
-                circle(app.width, app.height, 200f, 25)
+                circle(battleshipGame.width, battleshipGame.height, 200f, 25)
 
                 color = BLUE
-                circle(app.width, 0f, 200f, 25)
+                circle(battleshipGame.width, 0f, 200f, 25)
             }
         }
     }