+++ core/src/com/giancola/battleship/screens/PlacementScreen.kt
@@ -3,56 +3,70 @@ package com.giancola.battleship.screens
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.InputAdapter
 import com.badlogic.gdx.graphics.Texture
+import com.badlogic.gdx.math.GridPoint2
 import com.badlogic.gdx.scenes.scene2d.ui.Image
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
 import com.giancola.battleship.*
 import com.giancola.battleship.GameConstants.N_COLS
 import com.giancola.battleship.GameConstants.N_ROWS
-import com.giancola.battleship.actors.PlayerBoard
+import com.giancola.battleship.GameConstants.TILE_SIZE
+import com.giancola.battleship.actors.PlacementBoard
 import com.giancola.battleship.actors.ShipActor
+import com.giancola.battleship.actors.getBoundingBox
 import com.giancola.battleship.actors.overlaps
 import com.giancola.battleship.ui.Styles
 import ktx.actors.onClick
 import ktx.app.KtxScreen
+import java.lang.Math.round
+import kotlin.math.roundToInt
 
 
-class PlacementScreen(val app: App) : KtxScreen, InputAdapter() {
+class PlacementScreen(private val gameApp: BattleshipGame) : KtxScreen, InputAdapter() {
     private val playerData = PlayerData("Guerino", N_ROWS, N_COLS, ShipFactory.standardSetup)
     //private val playerData = PlayerData("Guerino", N_ROWS, N_COLS, ShipFactory.debugSetup)
 
     private val bkg: Image
-    val board: PlayerBoard
-    val confirmButton: TextButton
+    val board: PlacementBoard
     val ships: Map<Pair<ShipType, Int>, ShipActor>
+    private val confirmButton: TextButton
 
     init {
-        app.im.addProcessor(this)
+        gameApp.im.addProcessor(this)
 
         bkg = Image(Texture("placement_background.png"))
-        app.stg.addActor(bkg)
+        gameApp.stg.addActor(bkg)
 
-        board = PlayerBoard(this)
-        app.stg.addActor(board)
+        board = PlacementBoard(this)
+        gameApp.stg.addActor(board)
 
         confirmButton = TextButton("Confirm", Styles.buttonStyle)
         val r = LayoutConstants.standard2worldCoords(LayoutConstants.placementButton)
         confirmButton.setBounds(r.x, r.y, r.width, r.height)
         confirmButton.onClick {
-            Gdx.app.exit()
+            Gdx.app.log("Battleship", "Switching to combat screen")
+
+            updatePlayerData()
+
+            gameApp.stg.clear()
+            val combatScreen = CombatScreen(gameApp, playerData)
+            gameApp.addScreen(combatScreen)
+            gameApp.setScreen<CombatScreen>()
+
+            dispose()
         }
-        //confirmButton.isDisabled = true
-        confirmButton.isVisible = false
-        app.stg.addActor(confirmButton)
+        //confirmButton.isDisabled = true  //DEBUG
+        //confirmButton.isVisible = false
+        gameApp.stg.addActor(confirmButton)
 
 
         val mutableMap = mutableMapOf<Pair<ShipType, Int>, ShipActor>()
         for ((i, shipId) in this.playerData.shipPlacement.keys.withIndex()) {
             val shipActor = ShipActor(this, shipId)
 
-            val shipRect = LayoutConstants.standard2worldCoords(LayoutConstants.shipPositions[i])
+            val shipRect = LayoutConstants.standard2worldCoords(LayoutConstants.placementShipPositions[i])
             shipActor.setPosition(shipRect.x, shipRect.y)
 
-            app.stg.addActor(shipActor)
+            gameApp.stg.addActor(shipActor)
 
             mutableMap[shipId] = shipActor
         }
@@ -62,13 +76,13 @@ class PlacementScreen(val app: App) : KtxScreen, InputAdapter() {
     }
 
     override fun render(delta: Float) {
-        app.stg.act()
+        gameApp.stg.act()
 
-        app.stg.draw()
+        gameApp.stg.draw()
     }
 
     override fun dispose() {
-        app.im.removeProcessor(this)
+        gameApp.im.removeProcessor(this)
         super.dispose()
     }
 
@@ -88,6 +102,21 @@ class PlacementScreen(val app: App) : KtxScreen, InputAdapter() {
 
         //enable confirm button if everything ship is placed
         //screen.confirmButton.isDisabled = !allValid
-        confirmButton.isVisible = allValid
+        //confirmButton.isVisible = allValid
+    }
+
+    private fun updatePlayerData() {
+        for (ship in ships.values) {
+            val bbox = ship.getBoundingBox()
+            val iGridStart = ((bbox.x - board.x) / TILE_SIZE).roundToInt()
+            val jGridStart = ((bbox.y - board.y) / TILE_SIZE).roundToInt()
+            val iGridEnd = ((bbox.x + bbox.width - board.x) / TILE_SIZE).roundToInt() - 1
+            val jGridEnd = ((bbox.y + bbox.height - board.y) / TILE_SIZE).roundToInt() - 1
+
+            playerData.shipPlacement[ship.shipId] = Pair(
+                GridPoint2(iGridStart, jGridStart),
+                GridPoint2(iGridEnd, jGridEnd)
+            )
+        }
     }
 }