+++ desktop/src/com/giancola/battleship/DesktopLauncher.kt
@@ -2,8 +2,6 @@ package com.giancola.battleship
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration
-import com.giancola.battleship.App
-import com.giancola.battleship.GameConstants
 
 object DesktopLauncher {
     @JvmStatic
@@ -13,6 +11,6 @@ object DesktopLauncher {
         config.height = GameConstants.WORLD_HEIGHT.toInt()
         config.title = "Battleship"
 
-        LwjglApplication(App(), config)
+        LwjglApplication(BattleshipGame(), config)
     }
 }